+++ lib/executor/src/executors/map.rs
@@ -88,7 +88,7 @@ impl SubgraphExecutorMap {
             .pool_timer(TokioTimer::new())
             .pool_idle_timeout(config.traffic_shaping.all.pool_idle_timeout)
             .pool_max_idle_per_host(config.traffic_shaping.max_connections_per_host);
-        if config.traffic_shaping.all.http2_only {
+        if config.traffic_shaping.all.allow_only_http2 {
             client_builder.http2_only(true);
         }
         let client: HttpClient = client_builder.build(build_https_connector(
@@ -553,13 +553,13 @@ impl SubgraphExecutorMap {
         let pool_idle_timeout = subgraph_config
             .pool_idle_timeout
             .unwrap_or(self.config.traffic_shaping.all.pool_idle_timeout);
-        // Override client only if pool idle timeout is customized, TLS config is provided, or http2_only differs
-        let subgraph_http2_only = subgraph_config
-            .http2_only
-            .unwrap_or(self.config.traffic_shaping.all.http2_only);
+        // Override client only if pool idle timeout is customized, TLS config is provided, or allow_only_http2 differs
+        let subgraph_allow_only_http2 = subgraph_config
+            .allow_only_http2
+            .unwrap_or(self.config.traffic_shaping.all.allow_only_http2);
         if pool_idle_timeout != self.config.traffic_shaping.all.pool_idle_timeout
             || subgraph_config.tls.is_some()
-            || subgraph_http2_only != self.config.traffic_shaping.all.http2_only
+            || subgraph_allow_only_http2 != self.config.traffic_shaping.all.allow_only_http2
         {
             let tls_config = get_merged_tls_config(
                 self.config.traffic_shaping.all.tls.as_ref(),
@@ -570,7 +570,7 @@ impl SubgraphExecutorMap {
                 .pool_timer(TokioTimer::new())
                 .pool_idle_timeout(pool_idle_timeout)
                 .pool_max_idle_per_host(self.max_connections_per_host);
-            if subgraph_http2_only {
+            if subgraph_allow_only_http2 {
                 client_builder.http2_only(true);
             }
             config.client =
